+++ Tests/Functional/Traits/FalHelper.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace OliverKlee\Seminars\Tests\Functional\Traits;
+
+use Prophecy\Prophecy\ObjectProphecy;
+use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
+
+/**
+ * This trait provides methods useful when testing FAL output.
+ *
+ * @author Oliver Klee <typo3-coding@oliverklee.de>
+ */
+trait FalHelper
+{
+    /**
+     * Creates an admin BE user to allow FAL to access any file.
+     *
+     * This is necessary to as the nimut testing framework always runs with
+     * TYPO3_REQUESTTYPE_BE | TYPO3_REQUESTTYPE_CLI.
+     *
+     * @return void
+     */
+    private function provideAdminBackEndUserForFal()
+    {
+        /** @var BackendUserAuthentication|ObjectProphecy $backEndUserProphecy */
+        $backEndUserProphecy = $this->prophesize(BackendUserAuthentication::class);
+        $backEndUserProphecy->isAdmin()->willReturn(true);
+        $GLOBALS['BE_USER'] = $backEndUserProphecy->reveal();
+    }
+}