+++ Tests/Unit/Model/SpeakerTest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace OliverKlee\Seminars\Tests\Functional\Model;
+
+use Nimut\TestingFramework\TestCase\UnitTestCase;
+
+/**
+ * Test case.
+ *
+ * @author Oliver Klee <typo3-coding@oliverklee.de>
+ */
+class SpeakerTest extends UnitTestCase
+{
+    /**
+     * @var \Tx_Seminars_Model_Speaker
+     */
+    private $subject = null;
+
+    protected function setUp()
+    {
+        $this->subject = new \Tx_Seminars_Model_Speaker();
+    }
+
+    /**
+     * @test
+     */
+    public function hasImageWithoutImageReturnsFalse()
+    {
+        $this->subject->setData([]);
+
+        self::assertFalse($this->subject->hasImage());
+    }
+
+    /**
+     * @test
+     */
+    public function hasImageWithImageReturnsTrue()
+    {
+        $this->subject->setData(['image' => 1]);
+
+        self::assertTrue($this->subject->hasImage());
+    }
+}